refactor: update customer API imports
