+++ src/app/(private)/customers/[id]/page.tsx
@@ -17,9 +17,9 @@ import { useParams, useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { postCustomerStore } from '../../../../services/api/customerStore';
-import { putCustomerUpdate } from '../../../../services/api/customerUpdate';
-import { getCustomers } from '../../../../services/api/single-customer';
+import { getCustomers } from '../../../../services/api/customers/customer';
+import { postCustomerStore } from '../../../../services/api/customers/store';
+import { putCustomerUpdate } from '../../../../services/api/customers/update';
 import { PanelContactData } from './PanelContactData';
 import { PanelLocation } from './PanelLocation';
 import { PanelRegistrationData } from './PanelRegistrationData';
