+++ src/services/api/customers/customer.ts
@@ -1,4 +1,4 @@
-import { api } from '../api';
+import { api } from '..';
 
 type CustomerData = {
   action: string | null;
